refactor(core): rename css_parser and css_lexer files
Closes #9273
